Fix BG color of errorText and CI steps / stages (#2)
* Fix background color in error messages from common.inc.sh

Errors should be white text on red background; the bash syntax
was broken for the bg color. This is the fix.

* Merge all CI steps / stages in on bash script; bumps version to 1.0.2

Travis-CI does not have matrix support in stages. So this is a first
interration to get the jobs done - independend of the CI which is
running those steps.

In addition, it bumps the patch version to 1.0.2