WIP: Add GitHub actions workflows
This include the addition of the github actions and also refactoring the
ci.sh file which is adjusted to work with the GitHub actions.
As for specific events a new workflows needs to be defined, the Github
actions are splitted in different files and the ci.sh file takes a new
argument to act on a specific event. Prior it was all in one decided by
the environment variable of the travis branch.