Fix the image reference for matomo example/tests
The dockerfile under example/ is used for tests. As it always pulled the
latest tag the tests failed because it did not test the image which was
freshly built but it used the remote image which currently is in a not
well tested state. This change will make the example docker image use
the freshly built image which should be tested against.